// clients/generated/hxui/types/candidateStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum CandidateStatus {
  Active,
  Withdrawn,
  Winner,
  ClaimableWinner,
}

export type CandidateStatusArgs = CandidateStatus;

export function getCandidateStatusEncoder(): FixedSizeEncoder<CandidateStatusArgs> {
  return getEnumEncoder(CandidateStatus);
}

export function getCandidateStatusDecoder(): FixedSizeDecoder<CandidateStatus> {
  return getEnumDecoder(CandidateStatus);
}

export function getCandidateStatusCodec(): FixedSizeCodec<
  CandidateStatusArgs,
  CandidateStatus
> {
  return combineCodec(getCandidateStatusEncoder(), getCandidateStatusDecoder());
}